Move Captain data to separate namespace and split file
