+++ WoWsShipBuilder.DataStructures/Captain/SkillPosition.cs
@@ -1,0 +1,10 @@
+﻿namespace WoWsShipBuilder.DataStructures.Captain;
+
+// ReSharper disable UnusedAutoPropertyAccessor.Global
+// ReSharper disable CollectionNeverUpdated.Global
+public class SkillPosition
+{
+    public int Tier { get; set; }
+    public int XPosition { get; set; }
+    public ShipClass ShipClass { get; set; }
+}
